backend/requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
python-dotenv
# Si vas a usar Whisper local:
# openai-whisper
# torch
pydantic==2.6.3
openai==1.30.1
httpx==0.27.0
